requirements.txt
aiogram
aiohttp
beautifulsoup4
sqlite3
feedparser
python-dotenv
cachetools
tenacity
